Wait to save records with provisional foreign keys until they resolve
Whenever a field of type `key` is assigned a negative foreign key,
it subscribes to the repository waiting for that key to be resolved.
When the record with that key as its id is saved, the key is
translated into the positive-valued integer that's stored in the
server's database. Any foreign key pointing to the record is updated,
and any pending saves of the records with those foreign keys happen
automatically.